requirements.txt
Flask==3.0.0
rembg==2.0.53
Pillow==10.0.1
gunicorn==21.2.0
